+++ com.unity.cinemachine/Editor/Utility/SaveDuringPlay.cs
@@ -77,7 +77,7 @@ namespace SaveDuringPlay
             List<T> objectsInScene = new List<T>();
             foreach (T b in Resources.FindObjectsOfTypeAll<T>())
             {
-                if (b == null)
+                if (b == null || b.gameObject == null)
                     continue;   // object was deleted
                 GameObject go = b.gameObject;
                 if (go.hideFlags == HideFlags.NotEditable || go.hideFlags == HideFlags.HideAndDontSave)